fix method findByLogin; fix class AuthenticationFilter; fix class AuthenticationServiceImpl.
