+++ src/main/java/mate/academy/controller/car/DeleteCarController.java
@@ -1,26 +1,22 @@
-package mate.academy.controllers.drivers;
+package mate.academy.controller.car;
 
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.lib.Injector;
-import mate.academy.model.Car;
 import mate.academy.service.CarService;
 
-public class GetMyCurrentCarsController extends HttpServlet {
-    private static final String DRIVER_ID = "driver_id";
+public class DeleteCarController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("mate.academy");
     private final CarService carService = (CarService) injector.getInstance(CarService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long driverId = (Long) req.getSession().getAttribute(DRIVER_ID);
-        List<Car> cars = carService.getAllByDriver(driverId);
-        req.setAttribute("cars", cars);
-        req.getRequestDispatcher("/WEB-INF/views/cars/cars.jsp").forward(req, resp);
+        Long id = Long.valueOf(req.getParameter("id"));
+        carService.delete(id);
+        resp.sendRedirect(req.getContextPath() + "/cars/");
     }
 }
