+++ src/main/java/mate/academy/controller/driver/DeleteDriverController.java
@@ -1,4 +1,4 @@
-package mate.academy.controllers.cars;
+package mate.academy.controller.driver;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,17 +6,18 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.lib.Injector;
-import mate.academy.service.CarService;
+import mate.academy.service.DriverService;
 
-public class DeleteCarController extends HttpServlet {
+public class DeleteDriverController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("mate.academy");
-    private final CarService carService = (CarService) injector.getInstance(CarService.class);
+    private final DriverService driverService = (DriverService) injector
+            .getInstance(DriverService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         Long id = Long.valueOf(req.getParameter("id"));
-        carService.delete(id);
-        resp.sendRedirect(req.getContextPath() + "/cars/");
+        driverService.delete(id);
+        resp.sendRedirect(req.getContextPath() + "/drivers/");
     }
 }
