+++ src/main/java/mate/academy/controller/manufacturer/CreateManufacturerController.java
@@ -1,4 +1,4 @@
-package mate.academy.controllers.manufacturer;
+package mate.academy.controller.manufacturer;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
