--- src/main/java/mate/academy/controllers/drivers/DisplayAllDriversController.java
@@ -1,26 +1,0 @@
-package mate.academy.controllers.drivers;
-
-import java.io.IOException;
-import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import mate.academy.lib.Injector;
-import mate.academy.model.Driver;
-import mate.academy.service.DriverService;
-
-public class DisplayAllDriversController extends HttpServlet {
-    private static final Injector injector = Injector.getInstance("mate.academy");
-    private final DriverService driverService = (DriverService) injector
-            .getInstance(DriverService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-
-        List<Driver> allDrivers = driverService.getAll();
-        req.setAttribute("drivers", allDrivers);
-        req.getRequestDispatcher("/WEB-INF/views/drivers/drivers.jsp").forward(req, resp);
-    }
-}
