+++ src/main/java/mate/academy/security/AuthenticationServiceImpl.java
@@ -1,19 +1,19 @@
 package mate.academy.security;
 
+import mate.academy.dao.DriverDao;
 import mate.academy.exception.AuthenticationException;
 import mate.academy.lib.Dao;
 import mate.academy.lib.Inject;
 import mate.academy.model.Driver;
-import mate.academy.service.DriverService;
 
 @Dao
 public class AuthenticationServiceImpl implements AuthenticationService {
     @Inject
-    private DriverService driverService;
+    private DriverDao driverDao;
 
     @Override
     public Driver login(String login, String password) throws AuthenticationException {
-        Driver driverFromDB = driverService.findByLogin(login)
+        Driver driverFromDB = driverDao.findByLogin(login)
                 .orElseThrow(() -> new AuthenticationException("Incorrect login or password"));
 
         if (driverFromDB.getPassword().equals(password)) {
