+++ src/main/java/mate/academy/service/mpl/DriverServiceImpl.java
@@ -1,7 +1,6 @@
 package mate.academy.service.mpl;
 
 import java.util.List;
-import java.util.Optional;
 import mate.academy.dao.DriverDao;
 import mate.academy.lib.Inject;
 import mate.academy.lib.Service;
@@ -39,7 +38,7 @@ public class DriverServiceImpl implements DriverService {
     }
 
     @Override
-    public Optional<Driver> findByLogin(String login) {
-        return driverDao.findByLogin(login);
+    public Driver findByLogin(String login) {
+        return driverDao.findByLogin(login).get();
     }
 }
