+++ src/main/webapp/web.xml
@@ -5,7 +5,7 @@
          version="4.0">
     <servlet>
         <servlet-name>index</servlet-name>
-        <servlet-class>mate.academy.controllers.IndexController</servlet-class>
+        <servlet-class>mate.academy.controller.IndexController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>index</servlet-name>
@@ -14,7 +14,7 @@
 
     <servlet>
         <servlet-name>allDrivers</servlet-name>
-        <servlet-class>mate.academy.controllers.drivers.DisplayAllDriversController</servlet-class>
+        <servlet-class>mate.academy.controller.driver.DisplayAllDriversController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allDrivers</servlet-name>
@@ -23,7 +23,7 @@
 
     <servlet>
         <servlet-name>CreateManufacturer</servlet-name>
-        <servlet-class>mate.academy.controllers.manufacturer.CreateManufacturerController</servlet-class>
+        <servlet-class>mate.academy.controller.manufacturer.CreateManufacturerController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>CreateManufacturer</servlet-name>
@@ -32,7 +32,7 @@
 
     <servlet>
         <servlet-name>CreateDriver</servlet-name>
-        <servlet-class>mate.academy.controllers.drivers.CreateDriverController</servlet-class>
+        <servlet-class>mate.academy.controller.driver.CreateDriverController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>CreateDriver</servlet-name>
@@ -41,7 +41,7 @@
 
     <servlet>
         <servlet-name>CreateCar</servlet-name>
-        <servlet-class>mate.academy.controllers.cars.CreateCarController</servlet-class>
+        <servlet-class>mate.academy.controller.car.CreateCarController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>CreateCar</servlet-name>
@@ -50,7 +50,7 @@
 
     <servlet>
         <servlet-name>addDriverToCar</servlet-name>
-        <servlet-class>mate.academy.controllers.cars.AddDriverToCarController</servlet-class>
+        <servlet-class>mate.academy.controller.car.AddDriverToCarController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>addDriverToCar</servlet-name>
@@ -59,7 +59,7 @@
 
     <servlet>
         <servlet-name>deleteDriver</servlet-name>
-        <servlet-class>mate.academy.controllers.drivers.DeleteDriverController</servlet-class>
+        <servlet-class>mate.academy.controller.driver.DeleteDriverController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteDriver</servlet-name>
@@ -68,7 +68,7 @@
 
     <servlet>
         <servlet-name>allManufacturers</servlet-name>
-        <servlet-class>mate.academy.controllers.manufacturer.DisplayAllManufacturersController</servlet-class>
+        <servlet-class>mate.academy.controller.manufacturer.DisplayAllManufacturersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allManufacturers</servlet-name>
@@ -77,7 +77,7 @@
 
     <servlet>
         <servlet-name>deleteManufacturer</servlet-name>
-        <servlet-class>mate.academy.controllers.manufacturer.DeleteManufacturerController</servlet-class>
+        <servlet-class>mate.academy.controller.manufacturer.DeleteManufacturerController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteManufacturer</servlet-name>
@@ -86,7 +86,7 @@
 
     <servlet>
         <servlet-name>allMyCars</servlet-name>
-        <servlet-class>mate.academy.controllers.drivers.GetMyCurrentCarsController</servlet-class>
+        <servlet-class>mate.academy.controller.driver.GetMyCurrentCarsController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allMyCars</servlet-name>
@@ -95,7 +95,7 @@
 
     <servlet>
         <servlet-name>allCars</servlet-name>
-        <servlet-class>mate.academy.controllers.cars.DisplayAllCarsController</servlet-class>
+        <servlet-class>mate.academy.controller.car.DisplayAllCarsController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allCars</servlet-name>
@@ -104,7 +104,7 @@
 
     <servlet>
         <servlet-name>deleteCar</servlet-name>
-        <servlet-class>mate.academy.controllers.cars.DeleteCarController</servlet-class>
+        <servlet-class>mate.academy.controller.car.DeleteCarController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteCar</servlet-name>
@@ -113,7 +113,7 @@
 
     <servlet>
         <servlet-name>login</servlet-name>
-        <servlet-class>mate.academy.controllers.LoginController</servlet-class>
+        <servlet-class>mate.academy.controller.LoginController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>login</servlet-name>
